Create inCellType, update inXmlAttribVal
Query was recording negative numbers like "     1,234.56-" due to the type agnostic method being used to populate the worksheet and Lawson's funky number formatting. Added the inCellType subroutine to record values in the worksheet while also formatting the cell according to the column type. Also updated inXmlAttribVal to return the string value of an attribute for use by the new subroutine, so existing references had to be updated as well.